refactor(server/service): split LanternService/LanternServer + consolidate tests (refs #213)
- Extract LanternServer (Run/gracefulShutdown/lifecycle) from service.go into
  new server.go (rule 7: two unrelated public types per file).
- Fold test files per rule 1 (1:1 source<->test):
  * tombstone_clamp_test.go -> service_test.go
  * fake_backend_test.go Test funcs -> service_test.go (helper stays per rule 5)
  * replication_test.go (MutationLog tests) -> service_test.go
  * graceful_shutdown_test.go -> server_test.go
  * replication_service_test.go renamed -> replication_test.go
  * server_test.go (black-box) renamed -> server_external_test.go
  * service_bufconn_test.go renamed -> service_external_test.go
- AGENTS.md rule 2: add carveout naming the white-box/black-box pair
  xxx_test.go + xxx_external_test.go (Go forces them into separate files).
