+++ server/service/fake_backend_test.go
@@ -2,16 +2,12 @@ package service
 
 import (
 	"context"
-	"errors"
-	"testing"
 	"time"
 
 	"github.com/anaregdesign/lantern/core/cache/graph"
 	coregraph "github.com/anaregdesign/lantern/core/graph"
 	"github.com/anaregdesign/lantern/core/hlc"
 	pb "github.com/anaregdesign/lantern/pb/graph/v1"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 )
 
 // fakeBackend is a hand-rolled Backend stub exercised by tests that want to
@@ -297,77 +293,4 @@ func (f *fakeBackend) SnapshotEdges() []graph.SnapshotEdge[string] {
 		}
 	}
 	return out
-}
-
-func TestLanternService_FakeBackend_PutGetDelete(t *testing.T) {
-	fb := newFakeBackend()
-	svc := NewLanternService(fb)
-	ctx := context.Background()
-
-	v := &pb.Vertex{Key: "a", Value: &pb.Vertex_String_{String_: "alpha"}}
-	if _, err := svc.PutVertices(ctx, &pb.PutVerticesRequest{Vertices: []*pb.Vertex{v}}); err != nil {
-		t.Fatalf("PutVertices: %v", err)
-	}
-	if fb.putVerticesCalls != 1 {
-		t.Errorf("putVerticesCalls = %d, want 1", fb.putVerticesCalls)
-	}
-
-	resp, err := svc.GetVertex(ctx, &pb.GetVertexRequest{Key: "a"})
-	if err != nil {
-		t.Fatalf("GetVertex: %v", err)
-	}
-	if got := resp.Vertex.GetString_(); got != "alpha" {
-		t.Errorf("value = %q, want \"alpha\"", got)
-	}
-
-	if _, err := svc.DeleteVertices(ctx, &pb.DeleteVerticesRequest{Keys: []string{"a"}}); err != nil {
-		t.Fatalf("DeleteVertices: %v", err)
-	}
-	if fb.deleteVertices != 1 {
-		t.Errorf("deleteVertices = %d, want 1", fb.deleteVertices)
-	}
-}
-
-func TestLanternService_FakeBackend_Illuminate_PropagatesError(t *testing.T) {
-	fb := newFakeBackend()
-	fb.neighborErr = errors.New("simulated cache failure")
-	svc := NewLanternService(fb)
-
-	_, err := svc.Illuminate(context.Background(), &pb.IlluminateRequest{Seed: "a", Step: 1, K: 1})
-	if err == nil {
-		t.Fatal("expected error, got nil")
-	}
-	// status.FromContextError turns a generic non-context error into Unknown.
-	if st, _ := status.FromError(err); st.Code() != codes.Unknown {
-		t.Errorf("status code = %v, want Unknown", st.Code())
-	}
-}
-
-func TestLanternService_FakeBackend_PutAndDeleteEdge(t *testing.T) {
-	fb := newFakeBackend()
-	svc := NewLanternService(fb)
-	ctx := context.Background()
-
-	if _, err := svc.PutEdge(ctx, &pb.PutEdgeRequest{Edge: &pb.Edge{Tail: "t", Head: "h", Weight: 2.5}}); err != nil {
-		t.Fatalf("PutEdge: %v", err)
-	}
-	if fb.putEdgesCalls != 1 {
-		t.Errorf("putEdgesCalls = %d, want 1", fb.putEdgesCalls)
-	}
-
-	resp, err := svc.GetEdge(ctx, &pb.GetEdgeRequest{Tail: "t", Head: "h"})
-	if err != nil {
-		t.Fatalf("GetEdge: %v", err)
-	}
-	if resp.Edge.Weight != 2.5 {
-		t.Errorf("weight = %v, want 2.5", resp.Edge.Weight)
-	}
-
-	del, err := svc.DeleteEdge(ctx, &pb.DeleteEdgeRequest{Tail: "t", Head: "h"})
-	if err != nil {
-		t.Fatalf("DeleteEdge: %v", err)
-	}
-	if !del.Existed {
-		t.Error("Existed = false, want true")
-	}
 }
